+++ apps/web-antd/src/views/monitor/logininfor/index.vue
@@ -7,7 +7,6 @@ import { Page, useVbenModal, type VbenFormProps } from '@vben/common-ui';
 import { getVxePopupContainer } from '@vben/utils';
 
 import { Modal, Popconfirm, Space } from 'ant-design-vue';
-import dayjs from 'dayjs';
 
 import {
   tableCheckboxEvent,
@@ -22,7 +21,7 @@ import {
   loginInfoRemove,
   userUnlock,
 } from '#/api/monitor/logininfo';
-import { downloadExcel } from '#/utils/file/download';
+import { commonDownloadExcel } from '#/utils/file/download';
 import { confirmDeleteModal } from '#/utils/modal';
 
 import { columns, querySchema } from './data';
@@ -37,6 +36,14 @@ const formOptions: VbenFormProps = {
   },
   schema: querySchema(),
   wrapperClass: 'grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4',
+  // 日期选择格式化
+  fieldMappingTime: [
+    [
+      'dateTime',
+      ['params[beginTime]', 'params[endTime]'],
+      ['YYYY-MM-DD 00:00:00', 'YYYY-MM-DD 23:59:59'],
+    ],
+  ],
 };
 
 const gridOptions: VxeGridProps = {
@@ -55,20 +62,6 @@ const gridOptions: VxeGridProps = {
   proxyConfig: {
     ajax: {
       query: async ({ page }, formValues = {}) => {
-        // 区间选择器处理
-        if (formValues?.dateTime) {
-          formValues.params = {
-            beginTime: dayjs(formValues.dateTime[0]).format(
-              'YYYY-MM-DD 00:00:00',
-            ),
-            endTime: dayjs(formValues.dateTime[1]).format(
-              'YYYY-MM-DD 23:59:59',
-            ),
-          };
-          Reflect.deleteProperty(formValues, 'dateTime');
-        } else {
-          Reflect.deleteProperty(formValues, 'params');
-        }
         return await loginInfoList({
           pageNum: page.currentPage,
           pageSize: page.pageSize,
@@ -149,6 +142,17 @@ async function handleUnlock() {
   canUnlock.value = false;
   tableApi.grid.clearCheckboxRow();
 }
+
+function handleDownloadExcel() {
+  commonDownloadExcel(
+    loginInfoExport,
+    '登录日志',
+    tableApi.formApi.form.values,
+    {
+      fieldMappingTime: formOptions.fieldMappingTime,
+    },
+  );
+}
 </script>
 
 <template>
@@ -164,13 +168,7 @@ async function handleUnlock() {
           </a-button>
           <a-button
             v-access:code="['monitor:logininfor:export']"
-            @click="
-              downloadExcel(
-                loginInfoExport,
-                '登录日志',
-                tableApi.formApi.form.values,
-              )
-            "
+            @click="handleDownloadExcel"
           >
             {{ $t('pages.common.export') }}
           </a-button>