+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 **REFACTOR**
 
 - 使用VxeTable重构OAuth账号绑定列表(替代antdv的Table)
+- commonDownloadExcel方法 支持处理区间选择器字段导出excel
+
+**OTHERS**
+
+- 废弃downloadExcel方法 统一使用commonDownloadExcel方法
 
 # 1.1.0